refactor: Split system_tray.py into focused modules (543 → 12 lines)
Split system_tray.py (543 lines) into 4 focused modules under 300 lines each:

**Modules Created:**
- system_tray_modules/icon_manager.py (156 lines): Windows compatibility, icon lifecycle
- system_tray_modules/event_handlers.py (172 lines): Mouse/keyboard events, menu handling
- system_tray_modules/weather_formatter.py (257 lines): Weather data formatting, alerts
- system_tray_modules/__init__.py (52 lines): Main TaskBarIcon class combining components

**Key Changes:**
- Reduced main file from 543 → 12 lines (backward compatibility wrapper)
- Clear separation of concerns with focused responsibilities
- Maintained all existing functionality and API compatibility
- Updated test import paths for new module structure
- All modules well under 500-line limit (largest: 257 lines)

**Benefits:**
- Improved maintainability through logical code organization
- Easier testing and debugging of individual components
- Better code reusability with mixin-based architecture
- Preserved backward compatibility for existing imports

Follows established refactoring patterns with systematic approach targeting
200-300 lines per module while maintaining test coverage.
